split everything into separate files and confirm everything (besides createConfig) works
